% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputVerification.R
\name{check_equal_length_pair}
\alias{check_equal_length_pair}
\title{check_equal_length_pair}
\usage{
check_equal_length_pair(y, y_hat)
}
\arguments{
\item{y}{\itemize{
\item numeric vector
}}

\item{y_hat}{\itemize{
\item numeric vector
}}
}
\value{
An error in case of length mismatch, nothing otherwise
}
\description{
check_equal_length_pair
}
