% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputVerification.R
\name{check_equal_length_trio}
\alias{check_equal_length_trio}
\title{check_equal_length_trio}
\usage{
check_equal_length_trio(y, y_hat, y_hat_bench)
}
\arguments{
\item{y}{\itemize{
\item numeric vector
}}

\item{y_hat}{\itemize{
\item numeric vector
}}

\item{y_hat_bench}{\itemize{
\item numeric vector
}}
}
\value{
An error in case of any length mismatch among the three vectors, nothing otherwise
}
\description{
check_equal_length_trio
}
