% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputVerification.R
\name{check_variable}
\alias{check_variable}
\title{check_variable}
\usage{
check_variable(variable, variable_name)
}
\arguments{
\item{variable}{\itemize{
\item Variable to check
}}

\item{variable_name}{\itemize{
\item Variable name to be used for the error message
}}
}
\value{
An error in case of length or type error, nothing otherwise
}
\description{
Helper function to avoid code repetition to check type and length of inputs
}
