% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorMeasures.R
\name{GMRAE}
\alias{GMRAE}
\title{GMRAE - Geometric Mean Relative Absolute Error}
\usage{
GMRAE(y, y_hat, y_hat_bench, result_type = c("measure", "raw"))
}
\arguments{
\item{y}{\itemize{
\item True values of the time series
}}

\item{y_hat}{\itemize{
\item Forecast values of the time series
}}

\item{y_hat_bench}{\itemize{
\item Forecast values of the time series using the benchmark model
}}

\item{result_type}{\itemize{
\item String indicating whether the function should return the raw values ("raw"), or the computed measure ("measure")
}}
}
\value{
According to the value of \code{result_type}:
\itemize{
\item{\code{measure} -> }{GMRAE computed according to the formula}
\item{\code{raw} -> }{Vector containing one GMRAE for each time step}
}
}
\description{
GMRAE = \eqn{(1/n \prod{t=0}^n  r_t )^(1/n)}
}
\examples{
y <- AirPassengers
splitting_point <- round(2*length(y)/3)
y_train <- y[1:splitting_point]
h <- 5

y_test <- y[(splitting_point+1):(splitting_point+h)]
y_hat_naive <- rep(tail(y_train,1),h)
y_hat_average <-  rep(mean(y_train),h)
results <- GMRAE(y_test,y_hat_average,y_hat_naive)
}
