% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputVerification.R
\name{input_check_results_type}
\alias{input_check_results_type}
\title{input_check_results_type}
\usage{
input_check_results_type(result_type)
}
\arguments{
\item{result_type}{\itemize{
\item Value of results_type to check - string
}}
}
\value{
An error in case of length or type error, nothing otherwise
}
\description{
input_check_results_type
}
