% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputVerification.R
\name{input_check_y_yhat}
\alias{input_check_y_yhat}
\title{input_check_y_yhat}
\usage{
input_check_y_yhat(y, y_hat)
}
\arguments{
\item{y}{\itemize{
\item Value of y to check - numeric vector
}}

\item{y_hat}{\itemize{
\item Value of y_hat to check - numeric vector
}}
}
\value{
An error in case of length,type error or dimension mismatch, nothing otherwise
}
\description{
input_check_y_yhat
}
