% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputVerification.R
\name{input_check_y_yhat_ybench}
\alias{input_check_y_yhat_ybench}
\title{input_check_y_yhat_ybench}
\usage{
input_check_y_yhat_ybench(y, y_hat, y_hat_bench)
}
\arguments{
\item{y}{\itemize{
\item Value of y to check - numeric vector
}}

\item{y_hat}{\itemize{
\item Value of y_hat to check - numeric vector
}}

\item{y_hat_bench}{\itemize{
\item Value of y_hat_bench to check - numeric vector
}}
}
\value{
An error in case of length,type error or dimension mismatch, nothing otherwise
}
\description{
input_check_y_yhat_ybench
}
