% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorMeasures.R
\name{MASE}
\alias{MASE}
\title{MASE - Mean Absolute Scaled Error}
\usage{
MASE(y, y_hat, result_type = c("measure", "raw"))
}
\arguments{
\item{y}{\itemize{
\item True values of the time series
}}

\item{y_hat}{\itemize{
\item Forecasted values of the time series
}}

\item{result_type}{\itemize{
\item String indicating whether the function should return the raw values ("raw"), or the computed measure ("measure")
}}
}
\value{
According to the value of \code{result_type}:
\itemize{
\item{\code{measure} -> }{MASE computed according to the formula}
\item{\code{raw} -> }{Vector containing one MASE for each time step}
}
}
\description{
MASE = \eqn{1/T \sum_{t=1}^T ( ( e_t )/((T/(T-1)) \sum_{i=2}^T  Y_i-Y_{i-1}))}
}
\examples{
y <- AirPassengers
splitting_point <- round(2*length(y)/3)
y_train <- y[1:splitting_point]
h <- 5

y_test <- y[(splitting_point+1):(splitting_point+h)]
y_hat_naive <- rep(tail(y_train,1),h)
results <- MASE(y_test,y_hat_naive)
}
