% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorMeasures.R
\name{RMdSPE}
\alias{RMdSPE}
\title{RMdSPE - Root Median Squared Percentage Error}
\usage{
RMdSPE(y, y_hat)
}
\arguments{
\item{y}{\itemize{
\item True values of the time series
}}

\item{y_hat}{\itemize{
\item Forecasted values of the time series
}}
}
\value{
RMdSPE value
}
\description{
RMdSPE = \eqn{\sqrt{ median(100  ((y_t - y_hat_t)/(y_t))^2)}}
}
\examples{
y <- AirPassengers
splitting_point <- round(2*length(y)/3)
y_train <- y[1:splitting_point]
h <- 5
y_test <- y[(splitting_point+1):(splitting_point+h)]
y_hat_naive <- rep(tail(y_train,1),h)
result <- RMdSPE(y_test,y_hat_naive)
}
\references{
Hyndman, R. J., & Koehler, A. B. (2006). Another look at measures of forecast accuracy. International journal of forecasting, 22(4), 679-688.
}
