% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorMeasures.R
\name{RMSPE}
\alias{RMSPE}
\title{RMSPE - Root Mean Squared Percentage Error}
\usage{
RMSPE(y, y_hat, result_type = c("measure", "raw"))
}
\arguments{
\item{y}{\itemize{
\item True values of the time series
}}

\item{y_hat}{\itemize{
\item Forecasted values of the time series
}}

\item{result_type}{\itemize{
\item String indicating whether the function should return the raw values ("raw"), or the computed measure ("measure")
}}
}
\value{
According to the value of \code{result_type}:
\itemize{
\item{\code{measure} -> }{RMSPE computed according to the formula}
\item{\code{raw} -> }{Vector containing one RMSPE for each time step}
}
}
\description{
RMSPE = \eqn{\sqrt{1/n \sum_{t=0}^n (100  ((y_t - y_hat_t)/(y_t))^2}}
}
\examples{
y <- AirPassengers
splitting_point <- round(2*length(y)/3)
y_train <- y[1:splitting_point]
h <- 5
y_test <- y[(splitting_point+1):(splitting_point+h)]
y_hat_naive <- rep(tail(y_train,1),h)
result <- RMSPE(y_test,y_hat_naive)
}
\references{
Hyndman, R. J., & Koehler, A. B. (2006). Another look at measures of forecast accuracy. International journal of forecasting, 22(4), 679-688.
}
